Remove PostBreakBlockEvent.java, PostSetBlockEvent.java.
Rename new events.

Code cleanup.
