+++ src/main/java/net/minestom/server/event/player/PlayerBlockPlaceEvent.java
@@ -4,6 +4,7 @@ import net.minestom.server.coordinate.BlockVec;
 import net.minestom.server.coordinate.Point;
 import net.minestom.server.entity.Player;
 import net.minestom.server.entity.PlayerHand;
+import net.minestom.server.event.block.SetBlockEvent;
 import net.minestom.server.event.trait.BlockEvent;
 import net.minestom.server.event.trait.CancellableEvent;
 import net.minestom.server.event.trait.PlayerInstanceEvent;
@@ -14,7 +15,7 @@ import org.jetbrains.annotations.NotNull;
 /**
  * Called when a player tries placing a block.
  *
- * @deprecated Use {@link net.minestom.server.event.block.PostSetBlockEvent} or {@link net.minestom.server.event.block.PreSetBlockEvent}
+ * @deprecated Use {@link net.minestom.server.event.block.PostSetBlockEvent} or {@link SetBlockEvent}
  */
 @Deprecated()
 public class PlayerBlockPlaceEvent implements PlayerInstanceEvent, BlockEvent, CancellableEvent {
